完成

+++ Src/Server/GameServer/GameServer/Entities/Character.cs
@@ -107,7 +107,7 @@ namespace GameServer.Entities
                 //Log.InfoFormat("PostProcess>Team:characterID:{0}:{1} {2}<{3}", this.Id, this.Info.Name, TeamUpdateTS, this, Team.changeTime);
                 if (TeamUpdateTS < this.Team.changeTime)//时间戳小于队伍更新的时间
                 {
-                    TeamUpdateTS = Team.changeTime;
+                    TeamUpdateTS = (int)Team.changeTime;
                     this.Team.PostProcess(message);
                 }
             }
